feat: separate browser and node builds
