+++ lib/index.js
@@ -1,11 +1,5 @@
 "use strict";
 
-var _fs = require("fs");
-
-var _fs2 = _interopRequireDefault(_fs);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
 var tags = require("./tags.json");
 
 /*
@@ -297,77 +291,10 @@ var fromBuffer = function fromBuffer(buffer) {
   return data;
 };
 
-/**
- * @param file {String}
- * @returns {Object}
- * @example
- * var exif = sync("~/Picture/IMG_1981.JPG");
- * console.log(exif.createTime);
- */
-var sync = function sync(file) {
-  if (!file) {
-    throw new Error("File not found");
-  }
-
-  var buffer = _fs2.default.readFileSync(file);
-
-  return fromBuffer(buffer);
-};
-
-/**
- * @param file {String}
- * @param callback {Function}
- * @example
- * async("~/Picture/IMG_0707.JPG", (err, data) => {
- *     if(err) {
- *         console.log(err);
- *     }
- *     if(data) {
- *         console.log(data.ExifOffset.createTime);
- *     }
- * }
- */
-var async = function async(file, callback) {
-  var data = void 0;
-
-  new Promise(function (resolve, reject) {
-    if (!file) {
-      reject(new Error("❓File not found."));
-    }
-
-    _fs2.default.readFile(file, function (err, buffer) {
-      if (err) {
-        reject(err);
-      } else {
-        try {
-          if (isValid(buffer)) {
-            var buf = buffer.slice(SOIMarkerLength);
-
-            data = APPnHandler(buf);
-
-            resolve(data);
-          } else if (isTiff(buffer)) {
-            data = EXIFHandler(buffer, false);
-
-            resolve(data);
-          } else {
-            reject(new Error("😱Unsupport file type."));
-          }
-        } catch (e) {
-          reject(e);
-        }
-      }
-    });
-  }, function (error) {
-    callback(error, undefined);
-  }).then(function (d) {
-    callback(undefined, d);
-  }).catch(function (error) {
-    callback(error, undefined);
-  });
-};
-
+exports.isTiff = isTiff;
+exports.isValid = isValid;
 exports.fromBuffer = fromBuffer;
-exports.parse = async;
-exports.parseSync = sync;
+exports.APPnHandler = APPnHandler;
+exports.EXIFHandler = EXIFHandler;
+exports.SOIMarkerLength = SOIMarkerLength;
 //# sourceMappingURL=index.js.map
